+++ src/app/Components/demo.component.ts
@@ -26,14 +26,14 @@ export class DemoComponent
 	public openLightbox(config?: GalleryConfigInterface):void
 	{
 		this.lightboxService.open([
-			{source: 'https://image.shutterstock.com/z/stock-photo-two-beach-chairs-on-tropical-vacation-at-sea-1411290431.jpg', copyright: 'https://www.shutterstock.com'},
-			{source: 'https://image.shutterstock.com/z/stock-photo-soft-waves-with-foam-of-ocean-on-the-sandy-beach-background-1383401564.jpg', description: 'beach', copyright: 'https://www.shutterstock.com'},
-			{source: 'https://gallery.yopriceville.com/var/albums/Backgrounds/Background_Sea_Beach.jpg?m=1432232957', description: 'One huge image O.o', copyright: 'https://gallery.yopriceville.com/Backgrounds/Background_Sea_Beach'},
-			{source: 'assets/images/image1.jpg', copyright: 'unknown'},
-			{source: 'assets/images/image2.jpg', description: 'test', copyright: 'unknown'},
-			{source: 'assets/images/image3.jpg', description: 'test 2', copyright: 'unknown'},
-			{source: 'assets/images/image4.jpg', description: 'test 3', copyright: 'unknown'},
-			{source: 'assets/images/image5.jpg', copyright: 'unknown'},
+			{type: 'image', source: 'https://image.shutterstock.com/z/stock-photo-two-beach-chairs-on-tropical-vacation-at-sea-1411290431.jpg', copyright: 'https://www.shutterstock.com'},
+			{type: 'image', source: 'https://image.shutterstock.com/z/stock-photo-soft-waves-with-foam-of-ocean-on-the-sandy-beach-background-1383401564.jpg', description: 'beach', copyright: 'https://www.shutterstock.com'},
+			{type: 'image', source: 'https://gallery.yopriceville.com/var/albums/Backgrounds/Background_Sea_Beach.jpg?m=1432232957', description: 'One huge image O.o', copyright: 'https://gallery.yopriceville.com/Backgrounds/Background_Sea_Beach'},
+			{type: 'image', source: 'assets/images/image1.jpg', copyright: 'unknown'},
+			{type: 'image', source: 'assets/images/image2.jpg', description: 'test', copyright: 'unknown'},
+			{type: 'image', source: 'assets/images/image3.jpg', description: 'test 2', copyright: 'unknown'},
+			{type: 'image', source: 'assets/images/image4.jpg', description: 'test 3', copyright: 'unknown'},
+			{type: 'image', source: 'assets/images/image5.jpg', copyright: 'unknown'},
 		], config);
 
 		return;
@@ -42,15 +42,15 @@ export class DemoComponent
 	public openMixedLightbox(config?: GalleryConfigInterface):void
 	{
 		this.lightboxService.open([
-			{mp4Source: 'assets/videos/720p.mp4', description: 'test', copyright: 'unknown'},
-			{mp4Source: {
+			{type: 'video', mp4Source: 'assets/videos/720p.mp4', description: 'test', copyright: 'unknown'},
+			{type: 'video', mp4Source: {
 				240: 'assets/videos/240p.mp4',
 				480: 'assets/videos/480p.mp4',
 				720: 'assets/videos/720p.mp4',
-			}, description: 'test', copyright: 'unknown'},
-			{source: 'https://image.shutterstock.com/z/stock-photo-two-beach-chairs-on-tropical-vacation-at-sea-1411290431.jpg', copyright: 'https://www.shutterstock.com'},
-			{source: 'https://image.shutterstock.com/z/stock-photo-soft-waves-with-foam-of-ocean-on-the-sandy-beach-background-1383401564.jpg', description: 'beach', copyright: 'https://www.shutterstock.com'},
-			{source: 'https://gallery.yopriceville.com/var/albums/Backgrounds/Background_Sea_Beach.jpg?m=1432232957', description: 'One huge image O.o', copyright: 'https://gallery.yopriceville.com/Backgrounds/Background_Sea_Beach'},
+			}, description: 'test', copyright: 'unknown', resolution: {width: 1920, height: 1080}},
+			{type: 'image', source: 'https://image.shutterstock.com/z/stock-photo-two-beach-chairs-on-tropical-vacation-at-sea-1411290431.jpg', copyright: 'https://www.shutterstock.com'},
+			{type: 'image', source: 'https://image.shutterstock.com/z/stock-photo-soft-waves-with-foam-of-ocean-on-the-sandy-beach-background-1383401564.jpg', description: 'beach', copyright: 'https://www.shutterstock.com'},
+			{type: 'image', source: 'https://gallery.yopriceville.com/var/albums/Backgrounds/Background_Sea_Beach.jpg?m=1432232957', description: 'One huge image O.o', copyright: 'https://gallery.yopriceville.com/Backgrounds/Background_Sea_Beach'},
 		], config);
 
 		return;