+++ src/app/Services/lightbox.service.ts
@@ -25,6 +25,8 @@ export class LightboxService
 		enableImagePreloading: true,
 		startingIndex: 0,
 		enableAnimations: true,
+		ariaLabelNext: 'Next',
+		ariaLabelPrev: 'Previous',
 	};
 
 	constructor(